Scale from zero and to zero when required
Scaling to zero will pause a running task, scaling from zero will
either create a task when one does not exist - i.e. a restart of
the daemon, or unpause any task that is found in a paused state.

Tested on Ubuntu 16.04 x86_64 with scale-up flow after deleting
task and after pausing task with ctr.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>